requirements.txt
python-telegram-bot
pycurl
certifi
